improve object type annotation converter to properly handle methods
