+++ src/converters/convert_object_type_property.ts
@@ -1,15 +1,36 @@
 import {
+  isNullableTypeAnnotation,
   isTSFunctionType,
   ObjectTypeProperty,
   tsMethodSignature,
+  tsNullKeyword,
+  tsParenthesizedType,
   tsPropertySignature,
   tsTypeAnnotation,
+  tsUndefinedKeyword,
+  tsUnionType,
 } from '@babel/types';
 import { convertFlowType } from './convert_flow_type';
 import { baseNodeProps } from '../utils/baseNodeProps';
 
 export function convertObjectTypeProperty(property: ObjectTypeProperty) {
-  const tsType = convertFlowType(property.value);
+  let tsType;
+  if (!isNullableTypeAnnotation(property.value)) {
+    tsType = convertFlowType(property.value);
+  } else {
+    let tsValueT = convertFlowType(property.value.typeAnnotation);
+    if (isTSFunctionType(tsValueT)) {
+      tsValueT = tsParenthesizedType(tsValueT);
+    }
+    if (property.optional) {
+      // { key?: ?T } -> { key?: T | null }
+      tsType = tsUnionType([tsValueT, tsNullKeyword()]);
+    } else {
+      // { key: ?T } -> { key: T | null | undefined }
+      tsType = tsUnionType([tsValueT, tsUndefinedKeyword(), tsNullKeyword()]);
+    }
+  }
+
   if (property.method) {
     if (!isTSFunctionType(tsType)) {
       throw new Error('incorrect method declaration');
